Move container runner module to tests
