+++ tests/docker/test_docker.py
@@ -5,7 +5,7 @@ from unittest.mock import MagicMock, call
 
 import pytest
 
-from dj_link.docker import ContainerRunner
+from .docker import ContainerRunner
 
 
 def test_if_subclass_of_abstract_context_manager():
